refactor: move gsheets into its own package
