+++ packages/gsheets/graphql/index.js
@@ -1,0 +1,3 @@
+const { loadModule, addTypes } = require('apollo-modules-node')
+const { graphql: scalars } = require('@orbiting/backend-modules-scalars')
+module.exports = addTypes(loadModule(__dirname), [scalars])
