+++ packages/gsheets/graphql/resolvers/_queries/faqs.js
