+++ packages/gsheets/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  graphql: require('./graphql')
+}
