+++ servers/republik/package.json
@@ -16,7 +16,6 @@
     "emojione_minimal": "^2.1.2",
     "graphql": "^14.6.0",
     "graphql-fields": "^2.0.3",
-    "gsheets": "^2.0.0",
     "is-uuid": "^1.0.2",
     "isomorphic-unfetch": "^3.0.0",
     "lodash": "^4.17.11",
